new-credit-risk/requirements.txt
flask
azure-storage-blob
azure-identity
openai
pandas
scikit-learn
shap
python-docx
openpyxl
sentence-transformers
fastapi
uvicorn
joblib
requests
azure-ai-agents
azure-ai-projects
gunicorn
asyncio
semantic-kernel
json 
